[SECURITY-891] Fold tests back into normal files
